requirements.txt
# Core dependencies
anthropic>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Async support
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Logging and monitoring
structlog>=24.1.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
mypy>=1.7.0
ruff>=0.1.0
